[SOUND EDITOR][REFACTO] Move files to correct folder
