+++ src/modules/create/game-editor/editors/SoundEditor/InstrumentEditor.tsx
