+++ src/modules/create/game-editor/editors/SoundEditor/Note.ts
