+++ src/modules/create/game-editor/editors/SoundEditor/components/ProgressBar.tsx
