+++ src/modules/create/game-editor/editors/SoundEditor/constants/instruments.ts
