+++ src/modules/create/game-editor/editors/SoundEditor/instruments/contrabass.json
