+++ src/modules/create/game-editor/editors/SoundEditor/instruments/guitar.json
