+++ src/modules/create/game-editor/editors/SoundEditor/instruments/trumpet.json
