+++ src/modules/create/game-editor/editors/SoundEditor/styles/MusicGrid.styles.ts
