+++ src/modules/create/game-editor/editors/SoundEditor/styles/ProgressBar.styles.ts
